src/main/java/com/sos/validators/GenderValidator.java
package com.sos.validators;

import com.sos.models.enums.Gender;
import com.sos.validators.annotations.ValidGender;
import jakarta.validation.ConstraintValidator;
import jakarta.validation.ConstraintValidatorContext;

public class GenderValidator implements ConstraintValidator<ValidGender, String> {
    @Override
    public void initialize(ValidGender constraintAnnotation) {
        ConstraintValidator.super.initialize(constraintAnnotation);
    }

    @Override
    public boolean isValid(String value, ConstraintValidatorContext constraintValidatorContext) {
        return Gender.isValidGender(value);
    }
}
